+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - 改进运维指令。
   - com.dwarfeng.fdr.impl.service.telqos.RecordCommand。
+  - com.dwarfeng.fdr.impl.service.telqos.ViewCommand。
 
 #### Bug修复